+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bugfixes
 
+  - Fix inability to delete LCD ([#229](https://github.com/kizniche/mycodo/issues/229))
   - Fix measurements export
   - Fix display of BMP and BH1750 sensor measurements in sensor lists (graphs/export)